+++ cmd/kubeadm/app/cmd/phases/etcd_test.go
@@ -25,6 +25,8 @@ import (
 	cmdtestutil "k8s.io/kubernetes/cmd/kubeadm/test/cmd"
 )
 
+const phaseTestK8sVersion = "v1.11.0"
+
 func TestEtcdSubCommandsHasFlags(t *testing.T) {
 
 	subCmds := getEtcdSubCommands("", phaseTestK8sVersion)
