+++ cmd/kubeadm/app/cmd/phases/kubeconfig.go
@@ -18,34 +18,16 @@ package phases
 
 import (
 	"fmt"
-	"io"
 
 	"github.com/pkg/errors"
-	"github.com/spf13/cobra"
 	kubeadmapi "k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm"
-	kubeadmscheme "k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm/scheme"
-	kubeadmapiv1beta1 "k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm/v1beta1"
 	"k8s.io/kubernetes/cmd/kubeadm/app/cmd/phases/workflow"
-	cmdutil "k8s.io/kubernetes/cmd/kubeadm/app/cmd/util"
 	kubeadmconstants "k8s.io/kubernetes/cmd/kubeadm/app/constants"
 	kubeconfigphase "k8s.io/kubernetes/cmd/kubeadm/app/phases/kubeconfig"
 	"k8s.io/kubernetes/pkg/util/normalizer"
 )
 
 var (
-	kubeconfigLongDesc = normalizer.LongDesc(`
-	kubeconfig file utilities.
-	` + cmdutil.AlphaDisclaimer)
-
-	userKubeconfigLongDesc = normalizer.LongDesc(`
-	Outputs a kubeconfig file for an additional user.
-	` + cmdutil.AlphaDisclaimer)
-
-	userKubeconfigExample = normalizer.Examples(`
-	# Outputs a kubeconfig file for an additional user named foo
-	kubeadm alpha kubeconfig user --client-name=foo
-	`)
-
 	kubeconfigFilePhaseProperties = map[string]struct {
 		name  string
 		short string
@@ -147,60 +129,4 @@ func runKubeConfigFile(kubeConfigFileName string) func(workflow.RunData) error {
 		// creates the KubeConfig file (or use existing)
 		return kubeconfigphase.CreateKubeConfigFile(kubeConfigFileName, data.KubeConfigDir(), data.Cfg())
 	}
-}
-
-// NewCmdKubeConfig returns main command for kubeconfig phase
-func NewCmdKubeConfig(out io.Writer) *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "kubeconfig",
-		Short: "kubeconfig file utilities",
-		Long:  kubeconfigLongDesc,
-	}
-
-	cmd.AddCommand(NewCmdUserKubeConfig(out, kubeadmconstants.KubernetesDir, ""))
-	return cmd
-}
-
-// NewCmdUserKubeConfig returns sub commands for kubeconfig phase
-func NewCmdUserKubeConfig(out io.Writer, outDir, defaultKubernetesVersion string) *cobra.Command {
-
-	cfg := &kubeadmapiv1beta1.InitConfiguration{}
-
-	// Default values for the cobra help text
-	kubeadmscheme.Scheme.Default(cfg)
-
-	var cfgPath, token, clientName string
-	var organizations []string
-
-	// Creates the UX Command
-	cmd := &cobra.Command{
-		Use:     "user",
-		Short:   "Outputs a kubeconfig file for an additional user",
-		Long:    userKubeconfigLongDesc,
-		Example: userKubeconfigExample,
-		Run: runCmdPhase(func(outDir string, cfg *kubeadmapi.InitConfiguration) error {
-			if clientName == "" {
-				return errors.New("missing required argument --client-name")
-			}
-
-			// if the kubeconfig file for an additional user has to use a token, use it
-			if token != "" {
-				return kubeconfigphase.WriteKubeConfigWithToken(out, cfg, clientName, token)
-			}
-
-			// Otherwise, write a kubeconfig file with a generate client cert
-			return kubeconfigphase.WriteKubeConfigWithClientCert(out, cfg, clientName, organizations)
-		}, &outDir, &cfgPath, cfg, defaultKubernetesVersion),
-	}
-
-	// Add flags to the command
-	cmd.Flags().StringVar(&cfg.CertificatesDir, "cert-dir", cfg.CertificatesDir, "The path where certificates are stored")
-	cmd.Flags().StringVar(&cfg.APIEndpoint.AdvertiseAddress, "apiserver-advertise-address", cfg.APIEndpoint.AdvertiseAddress, "The IP address the API server is accessible on")
-	cmd.Flags().Int32Var(&cfg.APIEndpoint.BindPort, "apiserver-bind-port", cfg.APIEndpoint.BindPort, "The port the API server is accessible on")
-	cmd.Flags().StringVar(&outDir, "kubeconfig-dir", outDir, "The path where to save the kubeconfig file")
-	cmd.Flags().StringVar(&token, "token", token, "The token that should be used as the authentication mechanism for this kubeconfig, instead of client certificates")
-	cmd.Flags().StringVar(&clientName, "client-name", clientName, "The name of user. It will be used as the CN if client certificates are created")
-	cmd.Flags().StringSliceVar(&organizations, "org", organizations, "The orgnizations of the client certificate. It will be used as the O if client certificates are created")
-
-	return cmd
 }
